fix: 클라이언트컴포넌트에서 메타태그가 적용이 안되는 부분을 서버 컴포넌트로 변경시켜서 해결
